pyproject.toml
[tool.poetry]
name = "cupcake"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.15.0"
markdown = "^3.6"
django-filter = "^24.1"
psycopg2-binary = "^2.9.9"
requests = "^2.31.0"
django-cors-headers = "^4.3.0"
django-dbbackup = "^4.1.0"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
hiredis = "^2.3.2"
redis = "^5.0.3"
django-redis = "^5.4.0"
beautifulsoup4 = "^4.12.3"
channels = "^4.0.0"
channels-redis = {extras = ["cryptography"], version = "^4.2.0"}
uvicorn = {extras = ["standard"], version = "^0.29.0"}
django-rq = "^2.10.2"
ffmpeg-python = "^0.2.0"
numpy = "^1.26.4"
python-docx = "^1.1.0"
webvtt-py = "^0.4.6"
pytesseract = "^0.3.10"
drf-chunked-upload = "^0.6.0"
pronto = "^2.5.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
